Notebook refactoring
Extracted modules and created subdirectories
Fixed tests
Rename the stack to NotebookStack
