+++ backend/dataall/modules/notebooks/__init__.py
@@ -1,10 +1,7 @@
 """Contains the code related to SageMaker notebooks"""
-from dataall.db.api import TargetType
-from dataall.modules.notebooks import gql, models, cdk, services, permissions
+from dataall.modules.notebooks import gql, cdk, services
 
 # importing of the common code
 import dataall.modules.common.sagemaker as common
 
-__all__ = ["gql", "models", "cdk", "services", "permissions", "common"]
-
-TargetType("notebook", permissions.GET_NOTEBOOK, permissions.UPDATE_NOTEBOOK)
+__all__ = ["gql", "cdk", "services", "common"]
