+++ backend/dataall/modules/notebooks/aws/client.py
@@ -1,7 +1,7 @@
 import logging
 
 from dataall.aws.handlers.sts import SessionHelper
-from dataall.modules.notebooks.models import SagemakerNotebook
+from dataall.modules.notebooks.db.models import SagemakerNotebook
 from botocore.exceptions import ClientError
 
 logger = logging.getLogger(__name__)
