+++ backend/dataall/modules/notebooks/services/permissions.py
