+++ backend/dataall/modules/notebooks/services/services.py
@@ -8,7 +8,6 @@ from typing import List, Dict
 from dataall.api.Objects.Stack import stack_helper
 from dataall.core.context import get_context as context
 from dataall.core.environment.db.repositories import EnvironmentResourceRepository
-from dataall.core.environment.models import EnvironmentResource
 from dataall.db.api import (
     ResourcePolicy,
     Environment, KeyValueTag, Stack,
@@ -21,8 +20,8 @@ from dataall.utils.naming_convention import (
     NamingConventionPattern,
 )
 from dataall.utils.slugify import slugify
-from dataall.modules.notebooks.models import SagemakerNotebook
-from dataall.modules.notebooks import permissions
+from dataall.modules.notebooks.db.models import SagemakerNotebook
+from dataall.modules.notebooks.services import permissions
 from dataall.modules.common.sagemaker.permissions import MANAGE_NOTEBOOKS, CREATE_NOTEBOOK
 from dataall.core.permission_checker import has_resource_permission, has_tenant_permission, has_group_permission
