+++ tests/modules/notebooks/cdk/test_sagemaker_notebook_stack.py
@@ -3,13 +3,13 @@ import json
 import pytest
 from aws_cdk import App
 
-from dataall.modules.notebooks.cdk.stacks import SagemakerNotebook
+from dataall.modules.notebooks.cdk.stacks import NotebookStack
 
 
 @pytest.fixture(scope='function', autouse=True)
 def patch_methods(mocker, db, notebook, env, org):
     mocker.patch(
-        'dataall.modules.notebooks.cdk.stacks.SagemakerNotebook.get_engine',
+        'dataall.modules.notebooks.cdk.stacks.NotebookStack.get_engine',
         return_value=db
     )
     mocker.patch(
@@ -17,7 +17,7 @@ def patch_methods(mocker, db, notebook, env, org):
         return_value="dataall-pivot-role-name-pytest",
     )
     mocker.patch(
-        'dataall.modules.notebooks.cdk.stacks.SagemakerNotebook.get_target',
+        'dataall.modules.notebooks.cdk.stacks.NotebookStack.get_target',
         return_value=notebook,
     )
     mocker.patch(
@@ -40,7 +40,7 @@ def patch_methods(mocker, db, notebook, env, org):
 @pytest.fixture(scope='function', autouse=True)
 def template(notebook):
     app = App()
-    SagemakerNotebook(app, 'SagemakerNotebook', target_uri=notebook.notebookUri)
+    NotebookStack(app, 'SagemakerNotebook', target_uri=notebook.notebookUri)
     return json.dumps(app.synth().get_stack_by_name('SagemakerNotebook').template)
